get tests function outside of the sdk
